pyproject.toml
[project]
name = "parse-pdf-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy==2.3.5",
    "openai==2.8.1",
    "pandas==2.3.3",
    "pdfplumber==0.11.8",
    "pillow==12.0.0",
    "python-dotenv==1.2.1",
    "tiktoken==0.12.0",
]
